Removed python 2.7 compatibility

+++ example/arith_lib/test_Dec.py
@@ -16,14 +16,14 @@ class DecTest(TestCase):
     """ Decrementer unit test class """
 
     def bench(self, width, speed, nrsamples=0):
-        
+
         """ Decrementer test bench
 
         width -- decrementer bit width
         speed -- SLOW, MEDIUM or FAST
         nrsamples -- required number of random samples, or exhaustive
                      test if not set (default)
-                     
+
         """
 
         A = Signal(intbv(0))
@@ -36,7 +36,7 @@ class DecTest(TestCase):
         @instance
         def stimulus():
             if nrsamples:
-                vals = [long(random()*(2**width)) for i in range(nrsamples)]
+                vals = [int(random()*(2**width)) for i in range(nrsamples)]
             else:
                 vals = range(2**width)
             for i in vals:
@@ -51,18 +51,18 @@ class DecTest(TestCase):
 
     def testDecLargeSlow(self):
         Simulation(self.bench(width=39, speed=SLOW, nrsamples=16)).run()
-        
+
     def testDecSmallFast(self):
         Simulation(self.bench(width=8, speed=FAST)).run()
-        
+
     def testDecLargeFast(self):
         Simulation(self.bench(width=39, speed=FAST, nrsamples=16)).run()
-         
+
 
 if __name__ == "__main__":
     unittest.main()
-       
-                 
+
+
