+++ example/arith_lib/test_LeadZeroDet.py
@@ -16,14 +16,14 @@ class LeadZeroDetTest(TestCase):
     """ Leading zeroes detector unit test class """
 
     def bench(self, width, speed, nrsamples=0):
-        
+
         """ Leading zeroes detector test bench
 
         width -- decrementer bit width
         speed -- SLOW, MEDIUM or FAST
         nrsamples -- required number of random samples, or exhaustive
                      test if not set (default)
-                     
+
         """
 
         A = Signal(intbv(0))
@@ -36,7 +36,7 @@ class LeadZeroDetTest(TestCase):
         @instance
         def stimulus():
             if nrsamples:
-                vals = [long(random()*(2**width)) for i in range(nrsamples)]
+                vals = [int(random()*(2**width)) for i in range(nrsamples)]
             else:
                 vals = range(2**width)
             for i in vals:
@@ -51,18 +51,18 @@ class LeadZeroDetTest(TestCase):
 
     def testLeadZeroDetLargeSlow(self):
         Simulation(self.bench(width=39, speed=SLOW, nrsamples=16)).run()
-        
+
     def testLeadZeroDetSmallFast(self):
         Simulation(self.bench(width=8, speed=FAST)).run()
-        
+
     def testLeadZeroDetLargeFast(self):
         Simulation(self.bench(width=39, speed=FAST, nrsamples=16)).run()
-         
+
 
 if __name__ == "__main__":
     unittest.main()
-       
-                 
+
+
